requirements.txt
torchvision
numpy
opencv-python
matplotlib
pyyaml
selenium